+++ src/interfaces/IInterestRateModel.sol
@@ -3,5 +3,8 @@ pragma solidity ^0.8.0;
 
 interface IInterestRateModel {
     // gets borrow and supply interest rate per second
-    function getRatesPerSecondX96(uint cash, uint debt) external view returns (uint borrowRateX96, uint supplyRateX96);
+    function getRatesPerSecondX96(uint256 cash, uint256 debt)
+        external
+        view
+        returns (uint256 borrowRateX96, uint256 supplyRateX96);
 }
